refactor: Rename option to UIOption
